+++ settings/core.go
@@ -1,28 +1,22 @@
 package settings
 
 type Config struct {
-	CommentWidth               int
-	PlainHeadlines             bool
-	HighlightHeadlines         bool
-	HighlightComments          bool
-	EmojiSmileys               bool
-	DoNotMarkSubmissionsAsRead bool
-	HideIndentSymbol           bool
-	IndentationSymbol          string
-	DebugMode                  bool
-	EnableNerdFonts            bool
-	LesskeyPath                string
-	AutoExpandComments         bool
+	CommentWidth                int
+	DisableHeadlineHighlighting bool
+	DisableCommentHighlighting  bool
+	DisableEmojis               bool
+	DoNotMarkSubmissionsAsRead  bool
+	HideIndentSymbol            bool
+	IndentationSymbol           string
+	DebugMode                   bool
+	EnableNerdFonts             bool
+	LesskeyPath                 string
+	AutoExpandComments          bool
 }
 
 func Default() *Config {
 	return &Config{
-		CommentWidth:               70,
-		HighlightHeadlines:         true,
-		HighlightComments:          true,
-		EmojiSmileys:               true,
-		DoNotMarkSubmissionsAsRead: false,
-		HideIndentSymbol:           false,
-		IndentationSymbol:          " ▎",
+		CommentWidth:      70,
+		IndentationSymbol: " ▎",
 	}
 }